feat: Add fullscreen offset setting
This commit introduces a new setting to adjust the clock's position when the app is in fullscreen mode.  A submenu is added to allow users to specify an offset (in points) from the top-right corner of the screen.  The default offset is 0 (no offset). The setting is stored in UserDefaults under the key `kFullscreenOffsetX`.